rootfs/leaf-isle-bagger/requirements.txt
requests>=2.31
python-swiftclient